post-intex/comments/likes functionalities

+++ app/controllers/api/comments_controller.rb
@@ -1,38 +1,31 @@
 class Api::CommentsController < ApplicationController
-  def index
-    @comments = Comment.all
-    render :index
-  end
+    before_action :require_logged_in
 
-  def show
-    @comment = Comment.find(params[:id])
-    render :show
-  end
-  
-  def create
-    @comment = Comment.new(comment_params)
-    @comment.user_id = current_user.id
+    def create 
+        @comment = Comment.new(comment_params)
+        @comment.user_id = current_user.id 
 
-    if @comment.save
-      render :show, status: 200
-    else
-      render json: @comment.errors.full_messages, status: 422
+        if @comment.save 
+            render :show, status: 200
+        else
+            render json: @comment.errors.full_messages, status: 422
+        end
     end
-  end
 
-  def destroy
-    @comment = Comment.find(params[:id])
+    def destroy
+        @comment = Comment.find_by(id: params[:id])
 
-    if @comment.destroy 
-        render json: {}, status: 200 # or the show page of comment
-    else
-        render json: ["can't be processed"], status: 422
+        if @comment && @comment.user_id == current_user.id
+            if @comment.destroy 
+                render json: {}, status: 200
+            else
+                render json: ["can't be processed"], status: 422
+            end
+        end
     end
-  end
 
-  private
-
-  def comment_params
-    params.require(:comment).permit(:body, :user_id, :post_id)
-  end
+    private
+    def comment_params
+        params.require(:comment).permit(:post_id, :body)
+    end
 end
